# Optimize bytes_queue